Bugfixes in cocktail and bubble, changed everything from interface to long and also in the respectable classes of algorithms.

+++ src/HeapSort.java
@@ -8,10 +8,10 @@
 
 public class HeapSort extends Algorithm {
 
-    private int time = 0;
-    private int storage = 0;
-    private int comparisons = 0;
-    private int arrayAccess = 0;
+    private long time = 0;
+    private long storage = 0;
+    private long comparisons = 0;
+    private long arrayAccess = 0;
 
     /**
      * Method sorts array and switches the values in array
@@ -90,7 +90,7 @@ public class HeapSort extends Algorithm {
      * The comparisons that the Algorithm made in the array, for example is 1 > 2.
      */
     @Override
-    public int getComparison() {
+    public long getComparison() {
         return comparisons;
     }
 
@@ -100,7 +100,7 @@ public class HeapSort extends Algorithm {
      * The times the array was accessed through the Methods
      */
     @Override
-    public int getArrayAccess() {
+    public long getArrayAccess() {
         return arrayAccess;
     }
 
@@ -110,7 +110,7 @@ public class HeapSort extends Algorithm {
      * The amount of storage that has been used for this algorithm.
      */
     @Override
-    public int getStorage() {
+    public long getStorage() {
         return storage;
     }
